+++ src/SoilTracers.jl
@@ -3,5 +3,6 @@ __precompile__()
 module SoilTracers
 
 include("consts.jl")
+include("tracers/tracers.jl")
 
 end # module